+++ src/WebPanelManager.ts
@@ -32,6 +32,7 @@ import { view_str$compile$options, view_str$compile$storageLayout, view_str$env_
 import * as NodePath from 'path';
 import * as CmsisConfigParser from './CmsisConfigParser'
 import * as os from 'os'
+import * as platform from './Platform';
 
 let _instance: WebPanelManager;
 
@@ -99,7 +100,7 @@ export class WebPanelManager {
         panel.webview.html = htmlFile.Read()
             .replace(/"[\w\-\.\/]+?\.(?:css|js)"/ig, (str) => {
                 const fileName = str.substr(1, str.length - 2); // remove '"'
-                const absPath = NodePath.normalize(htmlFolder.path + NodePath.sep + fileName);
+                const absPath = File.normalize(htmlFolder.path + NodePath.sep + fileName);
                 return `"${panel.webview.asWebviewUri(vscode.Uri.file(absPath)).toString()}"`;
             });
 
@@ -138,25 +139,25 @@ export class WebPanelManager {
                 const fileSuffix = validationObj.fileMatch.replace('**/*', '');
                 if (dataFilePath.endsWith(fileSuffix)) {
                     const path = resManager.getAppRootFolder().path + File.sep + validationObj.url;
-                    return new File(NodePath.normalize(path));
+                    return new File(File.normalize(path));
                 }
             }
             throw new Error(`not found model file for '${dataFilePath}'`)
         }
 
         const envList: any[] = [
-            { name: '${TargetName}', desc: view_str$env_desc$project_name },
-            { name: '${ProjectRoot}', desc: view_str$env_desc$project_root },
-            { name: '${OutDir}', desc: view_str$env_desc$output_dir },
+            // unify_builder specific variables
             { name: '${BuilderFolder}', desc: view_str$env_desc$builer_folder },
-            { name: '${ToolchainRoot}', desc: view_str$env_desc$toolchain_root },
             { name: '${CompilerPrefix}', desc: view_str$env_desc$compiler_prefix },
             { name: '${CompilerFolder}', desc: view_str$env_desc$compiler_folder }
         ];
 
-        const prjEnv = project.getProjectEnv() || {};
+        const prjEnv = project.getProjectVariables();
         for (const key in prjEnv) {
-            envList.push({ name: `%${key}%`, desc: `${prjEnv[key]}` })
+            envList.push({
+                name: `\$\{${key}\}`,
+                desc: `${prjEnv[key]}`
+            })
         }
 
         /* prepare page-init event data */
@@ -220,7 +221,7 @@ export class WebPanelManager {
         panel.webview.html = htmlEntryFile.Read()
             .replace(/"[\w\-\.\/]+?\.(?:css|js)"/ig, (str) => {
                 const fileName = str.substr(1, str.length - 2); // remove '"'
-                const absPath = NodePath.normalize(htmlFolder.path + NodePath.sep + fileName);
+                const absPath = File.normalize(htmlFolder.path + NodePath.sep + fileName);
                 return `"${panel.webview.asWebviewUri(vscode.Uri.file(absPath)).toString()}"`;
             });
 
@@ -331,7 +332,7 @@ export class WebPanelManager {
         panel.webview.html = htmlEntryFile.Read()
             .replace(/"[\w\-\.\/]+?\.(?:css|js)"/ig, (str) => {
                 const fileName = str.substr(1, str.length - 2); // remove '"'
-                const absPath = NodePath.normalize(htmlFolder.path + NodePath.sep + fileName);
+                const absPath = File.normalize(htmlFolder.path + NodePath.sep + fileName);
                 return `"${panel.webview.asWebviewUri(vscode.Uri.file(absPath)).toString()}"`;
             });